ultimo

+++ app/Http/Controllers/UsersController.php
@@ -37,37 +37,43 @@ class UsersController extends Controller
     	while (PrincipalUser::where('id', $userId)->exists()) {
         	$userId = mt_rand(10000, 99999);
     	}
-
-        PrincipalUser::create([
-            'id' => $userId,
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-            'dep_id' => $request->id_depart,
-            'admin' => true,
-        ]);
-
-        PrincipalPermisos::updateOrCreate(['user_id' => $userId]);
-        
+        $instituto = $request->instituto;
         $depar = Departamento::find($request->id_depart);
-        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE'])) {
-            // Cambia la configuraciÃ³n de la base de datos
+        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE'])  && in_array($instituto, ['u', 'i', 'c']) ) {
             $databaseConnection = "departamento_" . strtolower($depar->sigla);
-            config(['database.default' => $databaseConnection]);
-            DB::reconnect();
-            $databaseName = config('database.connections.' . config('database.default') . '.database');
-            // Crea el usuario en la segunda base de datos con la misma ID
-            User::create([
-                'id' => $userId,
-                'name' => $request->name,
-                'email' => $request->email,
-                'password' => Hash::make($request->password),
-                'dep_id' => $request->id_depart,
-                'admin' => true,
-            ]);
-            // Crea el permiso relacionado en la segunda base de datos
-            Permiso::updateOrCreate(['user_id' => $userId]);
-            return back()->with('success', 'Usuario Creado Exitosamente');
+            if ($instituto !== 'u') {
+                $databaseConnection .= '_'.strtolower($instituto);
+            }
+            try {
+                DB::connection($databaseConnection)->getPdo();
+                config(['database.default' => $databaseConnection]);
+                DB::reconnect();
+                User::create([
+                    'id' => $userId,
+                    'name' => $request->name,
+                    'email' => $request->email,
+                    'password' => Hash::make($request->password),
+                    'dep_id' => $request->id_depart,
+                    'instituto' => $instituto,
+                    'admin' => true,
+                ]);
+                Permiso::updateOrCreate(['user_id' => $userId]);
+                ///En la base de datos principal
+                PrincipalUser::create([
+                    'id' => $userId,
+                    'name' => $request->name,
+                    'email' => $request->email,
+                    'password' => Hash::make($request->password),
+                    'dep_id' => $request->id_depart,
+                    'instituto' => $instituto,
+                    'admin' => true,
+                ]);
+                PrincipalPermisos::updateOrCreate(['user_id' => $userId]);
+                return back()->with('success', 'Usuario Creado Exitosamente');
+            } catch (\Exception $e) {
+                // La conexión no existe
+                return back()->with('success', 'La base de datos no existe');
+            }
         }
         return back()->with('success', 'Credenciales invÃ¡lidas');
     }
@@ -92,62 +98,68 @@ class UsersController extends Controller
         $user1->update();
         // Obtener los permisos seleccionados
         $permisosSeleccionados = $request->input('permisos', []);
-        
-        PrincipalPermisos::updateOrCreate(
-            ['user_id' => $id], // Buscar el permiso por el ID de usuario
-            [
-                'user_id' => $id,
-                'crear_user' => in_array('crear_user', $permisosSeleccionados),
-                'dar_baja_item' => in_array('dar_baja_item', $permisosSeleccionados),
-                'crear_item' => in_array('crear_item', $permisosSeleccionados),
-                'borrar_item' => in_array('borrar_item', $permisosSeleccionados),
-                'editar_item' => in_array('editar_item', $permisosSeleccionados),
-                'exportar' => in_array('exportar', $permisosSeleccionados),
-                'crear_area' => in_array('crear_area', $permisosSeleccionados),
-                'editar_area' => in_array('editar_area', $permisosSeleccionados),
-                'borrar_area' => in_array('borrar_area', $permisosSeleccionados),
-                'agregar_activo' => in_array('agregar_activo', $permisosSeleccionados),
-                'eliminar_activo' => in_array('eliminar_activo', $permisosSeleccionados),
-                'editar_activo' => in_array('editar_activo', $permisosSeleccionados),
-                'crear_material' => in_array('crear_material', $permisosSeleccionados),
-            ]
-        );
+        $instituto = $user1->instituto;
         $depar = Departamento::find($user1->dep_id);
-        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE'])) {
+        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE'])  && in_array($instituto, ['u', 'i', 'c']) ) {
             // Cambia la configuraciÃ³n de la base de datos
             $databaseConnection = "departamento_" . strtolower($depar->sigla);
-            config(['database.default' => $databaseConnection]);
-            DB::reconnect();
-            $databaseName = config('database.connections.' . config('database.default') . '.database');
-            // Crea el usuario en la segunda base de datos con la misma ID
-            $user = User::find($id);
-            $user->name = $request->name;
-            $user->email = $request->email;
-            $user->update();
-            // Guardar los permisos en la base de datos
-            Permiso::updateOrCreate(
-                ['user_id' => $id], // Buscar el permiso por el ID de usuario
-                [
-                    'user_id' => $id,
-                    'crear_user' => in_array('crear_user', $permisosSeleccionados),
-                    'dar_baja_item' => in_array('dar_baja_item', $permisosSeleccionados),
-                    'crear_item' => in_array('crear_item', $permisosSeleccionados),
-                    'borrar_item' => in_array('borrar_item', $permisosSeleccionados),
-                    'editar_item' => in_array('editar_item', $permisosSeleccionados),
-                    'exportar' => in_array('exportar', $permisosSeleccionados),
-                    'crear_area' => in_array('crear_area', $permisosSeleccionados),
-                    'editar_area' => in_array('editar_area', $permisosSeleccionados),
-                    'borrar_area' => in_array('borrar_area', $permisosSeleccionados),
-                    'agregar_activo' => in_array('agregar_activo', $permisosSeleccionados),
-                    'eliminar_activo' => in_array('eliminar_activo', $permisosSeleccionados),
-                    'editar_activo' => in_array('editar_activo', $permisosSeleccionados),
-                    'crear_material' => in_array('crear_material', $permisosSeleccionados),
-                ]
-            );
-            
-            return back()->with('success', 'Cuenta de usuario actualizado');
+            if ($instituto !== 'u') {
+                $databaseConnection .= '_'.strtolower($instituto);
+            }
+            try {
+                DB::connection($databaseConnection)->getPdo();
+                config(['database.default' => $databaseConnection]);
+                DB::reconnect();
+                // Crea el usuario en la segunda base de datos con la misma ID
+                $user = User::find($id);
+                $user->name = $request->name;
+                $user->email = $request->email;
+                $user->update();
+                // Guardar los permisos en la base de datos
+                Permiso::updateOrCreate(
+                    ['user_id' => $id], // Buscar el permiso por el ID de usuario
+                    [
+                        'user_id' => $id,
+                        'crear_user' => in_array('crear_user', $permisosSeleccionados),
+                        'dar_baja_item' => in_array('dar_baja_item', $permisosSeleccionados),
+                        'crear_item' => in_array('crear_item', $permisosSeleccionados),
+                        'borrar_item' => in_array('borrar_item', $permisosSeleccionados),
+                        'editar_item' => in_array('editar_item', $permisosSeleccionados),
+                        'exportar' => in_array('exportar', $permisosSeleccionados),
+                        'crear_area' => in_array('crear_area', $permisosSeleccionados),
+                        'editar_area' => in_array('editar_area', $permisosSeleccionados),
+                        'borrar_area' => in_array('borrar_area', $permisosSeleccionados),
+                        'agregar_activo' => in_array('agregar_activo', $permisosSeleccionados),
+                        'eliminar_activo' => in_array('eliminar_activo', $permisosSeleccionados),
+                        'editar_activo' => in_array('editar_activo', $permisosSeleccionados),
+                        'crear_material' => in_array('crear_material', $permisosSeleccionados),
+                    ]
+                );
+                PrincipalPermisos::updateOrCreate(
+                    ['user_id' => $id], // Buscar el permiso por el ID de usuario
+                    [
+                        'user_id' => $id,
+                        'crear_user' => in_array('crear_user', $permisosSeleccionados),
+                        'dar_baja_item' => in_array('dar_baja_item', $permisosSeleccionados),
+                        'crear_item' => in_array('crear_item', $permisosSeleccionados),
+                        'borrar_item' => in_array('borrar_item', $permisosSeleccionados),
+                        'editar_item' => in_array('editar_item', $permisosSeleccionados),
+                        'exportar' => in_array('exportar', $permisosSeleccionados),
+                        'crear_area' => in_array('crear_area', $permisosSeleccionados),
+                        'editar_area' => in_array('editar_area', $permisosSeleccionados),
+                        'borrar_area' => in_array('borrar_area', $permisosSeleccionados),
+                        'agregar_activo' => in_array('agregar_activo', $permisosSeleccionados),
+                        'eliminar_activo' => in_array('eliminar_activo', $permisosSeleccionados),
+                        'editar_activo' => in_array('editar_activo', $permisosSeleccionados),
+                        'crear_material' => in_array('crear_material', $permisosSeleccionados),
+                    ]
+                );
+                return back()->with('success', 'Cuenta de usuario actualizado');
+            } catch (\Exception $e) {
+                // La conexión no existe
+                return back()->with('success', 'La base de datos no existe');
+            }
         }
-
         return redirect('admin/users')->with('success', 'Cuenta de usuario actualizado');
     }
 
@@ -158,17 +170,24 @@ class UsersController extends Controller
         $user1->update();
         $depar = Departamento::find($user1->dep_id);
         $userAuth = PrincipalUser::find(auth()->user()->id);
-        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE'])) {
+        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE']) && in_array($user1->instituto, ['u', 'i', 'c'])) {
             $databaseConnection = "departamento_" . strtolower($depar->sigla);
-            config(['database.default' => $databaseConnection]);
-            DB::reconnect();
-            $databaseName = config('database.connections.' . config('database.default') . '.database');
-            // Crea el usuario en la segunda base de datos con la misma ID
-            $user = User::find($id);
-            $user->estado = 'I';
-            $user->admin = false;
-            $user->update();
-            return back()->with('success', 'Se dio de baja la Cuenta');
+            if ($user1->instituto !== 'u') {
+                $databaseConnection .= '_'.strtolower($user1->instituto);
+            }
+            try {
+                DB::connection($databaseConnection)->getPdo();
+                config(['database.default' => $databaseConnection]);
+                DB::reconnect();
+                $user = User::find($id);
+                $user->estado = 'I';
+                $user->admin = false;
+                $user->update();
+                return back()->with('success', 'Se dio de baja la Cuenta');
+            } catch (\Exception $e) {
+                // La conexión no existe
+                return back()->with('success', 'La base de datos no existe');
+            }
         }
         return $userAuth->tipo_user == 1 ? back()->with('success', 'Se dio de baja la Cuenta') : redirect('/admin/error');
     }
@@ -180,17 +199,26 @@ class UsersController extends Controller
         $user1->update();
         $depar = Departamento::find($user1->dep_id);
         $userAuth = PrincipalUser::find(auth()->user()->id);
-        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE'])) {
+        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE']) && in_array($user1->instituto, ['u', 'i', 'c'])) {
             $databaseConnection = "departamento_" . strtolower($depar->sigla);
-            config(['database.default' => $databaseConnection]);
-            DB::reconnect();
-            // Crea el usuario en la segunda base de datos con la misma ID
-            $user = User::find($id);
-            $user->estado = 'A';
-            $user->admin = true;
-            $user->update();
-            $userAuth = User::find(auth()->user()->id);
-            return back()->with('success', 'Se activo la Cuenta');
+            if ($user1->instituto !== 'u') {
+                $databaseConnection .= '_'.strtolower($user1->instituto);
+            }
+            try {
+                DB::connection($databaseConnection)->getPdo();
+                config(['database.default' => $databaseConnection]);
+                DB::reconnect();
+                // Crea el usuario en la segunda base de datos con la misma ID
+                $user = User::find($id);
+                $user->estado = 'A';
+                $user->admin = true;
+                $user->update();
+                $userAuth = User::find(auth()->user()->id);
+                return back()->with('success', 'Se activo la Cuenta');
+            } catch (\Exception $e) {
+                // La conexión no existe
+                return back()->with('success', 'La base de datos no existe');
+            }
         }
         return $userAuth->tipo_user == 1 ? back()->with('success', 'Se activo la Cuenta') : redirect('/admin/error');
     }
@@ -206,14 +234,23 @@ class UsersController extends Controller
         $user1->update();
 
         $depar = Departamento::find($user1->dep_id);
-        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE'])) {
+        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE']) && in_array($user1->instituto, ['u', 'i', 'c'])) {
             $databaseConnection = "departamento_" . strtolower($depar->sigla);
-            config(['database.default' => $databaseConnection]);
-            DB::reconnect();
-            $user = User::find($id);
-            $user->password = Hash::make($request->password);
-            $user->update();
-            return redirect('admin/users')->with('success', 'Se restablecio la contraseña');
+            if ($user1->instituto !== 'u') {
+                $databaseConnection .= '_'.strtolower($user1->instituto);
+            }
+            try {
+                DB::connection($databaseConnection)->getPdo();
+                config(['database.default' => $databaseConnection]);
+                DB::reconnect();
+                $user = User::find($id);
+                $user->password = Hash::make($request->password);
+                $user->update();
+                return redirect('admin/users')->with('success', 'Se restablecio la contraseña');
+            } catch (\Exception $e) {
+                // La conexión no existe
+                return back()->with('success', 'La base de datos no existe');
+            }
         }
         return redirect('admin/users')->with('success', 'Se restablecio la contraseña');
     }
@@ -232,17 +269,26 @@ class UsersController extends Controller
         }
         $userAuth = User::find(auth()->user()->id);
         $depar = Departamento::find($userP->dep_id);
-        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE'])) {
+        if (in_array($depar->sigla, ['PO', 'SU', 'TJ', 'LP', 'CB', 'OR', 'SC', 'PA', 'BE']) && in_array($userP->instituto, ['u', 'i', 'c'])) {
             $databaseConnection = "departamento_" . strtolower($depar->sigla);
-            config(['database.default' => $databaseConnection]);
-            DB::reconnect();
-            $user = User::find($id);
-            if ($user) {
-                $permiso = Permiso::where('user_id', $user->id);
-                $permiso->delete();
-                $user->delete();
+            if ($userP->instituto !== 'u') {
+                $databaseConnection .= '_'.strtolower($userP->instituto);
             }
-            return $userAuth->tipo_user == 1 ? back()->with('success', 'Se elimino la cuenta') : redirect('/admin/error');
+            try {
+                DB::connection($databaseConnection)->getPdo();
+                config(['database.default' => $databaseConnection]);
+                DB::reconnect();
+                $user = User::find($id);
+                if ($user) {
+                    $permiso = Permiso::where('user_id', $user->id);
+                    $permiso->delete();
+                    $user->delete();
+                }
+                return $userAuth->tipo_user == 1 ? back()->with('success', 'Se elimino la cuenta') : redirect('/admin/error');
+            } catch (\Exception $e) {
+                // La conexión no existe
+                return back()->with('success', 'La base de datos no existe');
+            }
         }
         return $userAuth->tipo_user == 1 ? back()->with('success', 'Se elimino la cuenta') : redirect('/admin/error');
     }
